teste_projeto_sql/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
pytest
pytest-asyncio
python-dotenv
